refactor(desktop): split backend-provider commands out of agents.rs
The async conversions pushed agents.rs to 1464 lines, over its file-size
guard ceiling (1437). Rather than bump the limit, extract the two
self-contained backend-provider commands (`discover_backend_providers`,
`probe_backend_provider`) into `commands/agent_providers.rs`, matching the
existing `agent_discovery`/`agent_models`/`agent_config` split pattern. This
brings agents.rs to 1420 lines (under the ceiling) with no logic change —
the commands are glob-re-exported through `commands::*`, so the Tauri
invoke-handler registration is unchanged.
