+++ desktop/src-tauri/src/commands/mod.rs
@@ -1,6 +1,7 @@
 mod agent_config;
 mod agent_discovery;
 mod agent_models;
+mod agent_providers;
 mod agent_settings;
 mod agents;
 mod canvas;
@@ -34,6 +35,7 @@ mod workspace;
 pub use agent_config::*;
 pub use agent_discovery::*;
 pub use agent_models::*;
+pub use agent_providers::*;
 pub use agent_settings::*;
 pub use agents::*;
 pub use canvas::*;
